Add files via upload
Plot multiple data sets on a single waveform chart and customize the chart view. Modify the VI from Exercise 3-3 to plot both the current temperature and the running average on the same chart. In addition, allow the user to examine a portion of the plot while the data is being acquired.