Add files via upload
You created a VI where a user inputs a temperature, a maximum temperature, and a minimum
temperature. A warning string generates depending on the relationship of the given inputs. However, a
situation could occur that causes the VI to work incorrectly. The user could enter a maximum
temperature that is less than the minimum temperature. Modify the VI to generate a different string to
alert the user to the error: Upper Limit < Lower Limit. Set the Warning? indicator to True to
indicate the error.