Add files via upload
In the Determine Warnings VI from Exercise 3-5, you used the Select function to pass a string based on a
decision. Revise this block diagram to use Case structures to complete the same purpose.

As part of the challenge, no implementation instructions are given for this exercise. Start from the VI
downloaded from Blackboard.