Add files via upload
You have a VI that determines whether a Heatstroke Warning or a Freeze Warning has occurred, based
on temperature input. You must expand this VI so that it also determines whether a High Wind Warning
has occurred based on a wind speed reading and a maximum wind speed setting. The warnings must be
displayed as a single string. For example, if a Heatstroke Warning and a High Wind Warning has
occurred, the string should read: Heatstroke and High Wind Warning.